refactor/INT-34 refactor package - исправил положение папки entities
