+++ game/src/test/java/io/deeplay/camp/UnitTest.java
@@ -1,3 +1,9 @@
+package io.deeplay.camp;
+
+import io.deeplay.camp.entities.Archer;
+import io.deeplay.camp.entities.Healer;
+import io.deeplay.camp.entities.Knight;
+import io.deeplay.camp.entities.Mage;
 import io.deeplay.entities.*;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
